refactor: 에러 처리 리팩토링
